Create ds24
Data Source 24 csv file. First column is the main developer, the columns following in each row are developers that have worked on the same project as the main developer. If a developer appears as a secondary developer (dev2) in a row, the main developer (dev1) of that row will not appear as a secondary developer in the row where dev2 is the main developer.